refactor(editor tiles): Creacion de carpetas para separar e implementar el MVC
