+++ maps/combined_map.overlay.json
@@ -984,6 +984,7 @@
     "wal",
     "fl0",
     "fl0",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
@@ -992,8 +993,7 @@
     "fl0",
     "fl0",
     "fl0",
-    "fl0",
-    "fl0",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
@@ -1115,14 +1115,136 @@
     "fl0",
     "fl0",
     "fl0",
-    "fl0",
-    "fl0",
+    "",
+    "",
     "fl0",
     "fl0",
     "fl0",
     "fl5",
     "fl5",
     "fl5",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    ""
+  ],
+  [
+    "wal",
+    "wal",
+    "wal",
+    "wal",
+    "wal",
+    "wal",
+    "fl0",
+    "fl3",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl5",
+    "fl5",
+    "fl2",
     "fl2",
     "fl2",
     "fl2",
@@ -1236,130 +1358,8 @@
     "fl0",
     "fl0",
     "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl5",
-    "fl5",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    ""
-  ],
-  [
-    "wal",
-    "wal",
-    "wal",
-    "wal",
-    "wal",
-    "wal",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
+    "fl1",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
@@ -1471,6 +1471,7 @@
     "wal",
     "wal",
     "fl0",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
@@ -1480,8 +1481,7 @@
     "fl0",
     "fl0",
     "fl0",
-    "fl0",
-    "fl0",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
@@ -1593,7 +1593,7 @@
     "wal",
     "wal",
     "fl0",
-    "fl0",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
@@ -1715,7 +1715,7 @@
     "",
     "wal",
     "fl0",
-    "fl0",
+    "fl1",
     "fl0",
     "fl0",
     "fl0",
